Expand error message to point to alternatives (#3)
This expands the error message to 1) mention how to still read the file
if you know it is trusted, and 2) point users to use ExtensionType
instead of PyExtensionType.
(similarly as I did for pandas. Although if you actively installed
pyarrow_hotfix, you are probably already aware of it and it's less
important to have this expanded message).