Don't fail if pyarrow is not installed. (#6)
Since this doesn't directly depend upon pyarrow, it is possible that
some end-user code does `import pyarrow_hotfix` as a defense mechanism,
despite not needing `pyarrow_hotfix` itself. If this happens, it should
not be a failure.

---

This is a simple proposal to take or leave; it also may be reasonable to
add a warning that Arrow isn’t installed.